requirements.txt
numpy==2.3.3 
pandas==2.3.3
matplotlib==3.10.6
seaborn==0.13.2
pyyaml==6.0.3
scipy==1.16.2
statsmodels==0.14.5
# pandasai==3.0.0
# pandasai-litellm==0.0.1